% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrapper.R
\name{check_lat_lon}
\alias{check_lat_lon}
\title{Internal helper: validate latitude / longitude}
\usage{
check_lat_lon(latitude, longitude)
}
\arguments{
\item{latitude}{numeric scalar}

\item{longitude}{numeric scalar}
}
\value{
NULL (throws error if invalid)
}
\description{
Internal helper: validate latitude / longitude
}
\keyword{internal}
