% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/forest_comparison.R
\name{compare_forecasts}
\alias{compare_forecasts}
\title{Create a forecast_comparison object}
\usage{
compare_forecasts(locations, days = 3)

compare_forecasts(locations, days = 3)
}
\arguments{
\item{locations}{A character vector of city names}

\item{days}{Number of forecast days}
}
\value{
An object of class forecast_comparison

An object of class forecast_comparison
}
\description{
Create a forecast_comparison object

Create forecast_comparison object
}
