% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_forecast_result}
\alias{create_forecast_result}
\title{Create a forecast_result object}
\usage{
create_forecast_result(location, forecast_data)
}
\arguments{
\item{location}{Name of location (string)}

\item{forecast_data}{A data.frame with date, temp_max, temp_min, precipitation}
}
\value{
An object of class forecast_result
}
\description{
Create a forecast_result object
}
