% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forecast.R
\name{get_forecast_by_location}
\alias{get_forecast_by_location}
\title{Get Forecast Using a Place Name}
\usage{
get_forecast_by_location(location, days = 1)
}
\arguments{
\item{location}{A city or town name}

\item{days}{Number of forecast days}
}
\value{
A data.frame with forecast data
}
\description{
Get Forecast Using a Place Name
}
