% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrapper.R
\name{get_forecast_raw}
\alias{get_forecast_raw}
\title{Internal: call Open-Meteo forecast endpoint and return raw httr2 response}
\usage{
get_forecast_raw(
  latitude,
  longitude,
  start_date,
  end_date,
  daily_variables = c(),
  hourly_variables = c()
)
}
\arguments{
\item{latitude}{numeric scalar}

\item{longitude}{numeric scalar}

\item{start_date}{Date/string}

\item{end_date}{Date/string}

\item{daily_variables}{character vector}

\item{hourly_variables}{character vector}
}
\value{
httr2_response
}
\description{
Internal: call Open-Meteo forecast endpoint and return raw httr2 response
}
\keyword{internal}
