% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/plot_forecast.R
\name{plot.forecast_result}
\alias{plot.forecast_result}
\title{Plot method for forecast_result}
\usage{
\method{plot}{forecast_result}(x, ...)

\method{plot}{forecast_result}(x, ...)
}
\arguments{
\item{x}{A forecast_result object}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plot method for forecast_result
}
