% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plot Weather Forecast}
\usage{
plot_forecast(forecast_df)
}
\arguments{
\item{forecast_df}{A data frame returned from \code{get_forecast_by_location()} or \code{get_forecast()}}
}
\value{
A ggplot object showing temperature and precipitation
}
\description{
Plot Weather Forecast
}
