% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_comparison.R
\name{print.forecast_comparison}
\alias{print.forecast_comparison}
\title{Print method for forecast_comparison}
\usage{
\method{print}{forecast_comparison}(x, ...)
}
\arguments{
\item{x}{forecast_comparison object}
}
\description{
Print method for forecast_comparison
}
