% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.forecast_result}
\alias{print.forecast_result}
\title{Print method for forecast_result}
\usage{
\method{print}{forecast_result}(x, ...)
}
\arguments{
\item{x}{A forecast_result object}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print method for forecast_result
}
