% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/forest_comparison.R
\name{summary.forecast_comparison}
\alias{summary.forecast_comparison}
\title{Summary method for forecast_comparison}
\usage{
\method{summary}{forecast_comparison}(object, ...)

\method{summary}{forecast_comparison}(object, ...)
}
\arguments{
\item{object}{forecast_comparison object}
}
\description{
Summary method for forecast_comparison

Summary method for forecast_comparison
}
