% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{summary.forecast_result}
\alias{summary.forecast_result}
\title{Summary method for forecast_result}
\usage{
\method{summary}{forecast_result}(object, ...)
}
\arguments{
\item{object}{A forecast_result object}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary method for forecast_result
}
