% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forecast.R
\name{get_forecast}
\alias{get_forecast}
\title{Get weather forecast by coordinates}
\usage{
get_forecast(lat, lon, days = 1)
}
\arguments{
\item{lat}{Latitude}

\item{lon}{Longitude}

\item{days}{Number of forecast days}
}
\value{
An object of class forecast_result
}
\description{
Get weather forecast by coordinates
}
