% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forecastsxy.R
\name{get_forecast}
\alias{get_forecast}
\title{Get Forecast by Latitude and Longitude}
\usage{
get_forecast(lat, lon, days = 1)
}
\arguments{
\item{lat}{Latitude (e.g., 27.7 for Kathmandu)}

\item{lon}{Longitude (e.g., 85.3 for Kathmandu)}

\item{days}{Number of forecast days (default = 1)}
}
\value{
A data.frame with daily forecast (temperature and precipitation)
}
\description{
Get Forecast by Latitude and Longitude
}
