% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forecastsxy.R
\name{get_forecast_by_location}
\alias{get_forecast_by_location}
\title{Get forecast by place name}
\usage{
get_forecast_by_location(location, days = 1)
}
\arguments{
\item{location}{A city or town name}

\item{days}{Number of forecast days}
}
\value{
An object of class forecast_result
}
\description{
Get forecast by place name
}
