% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_comparison.R
\name{plot.forecast_comparison}
\alias{plot.forecast_comparison}
\title{Plot method for forecast_comparison}
\usage{
\method{plot}{forecast_comparison}(x, ...)
}
\arguments{
\item{x}{forecast_comparison object}
}
\description{
Plot method for forecast_comparison
}
