% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plot.forecast_result}
\alias{plot.forecast_result}
\title{Plot method for forecast_result}
\usage{
\method{plot}{forecast_result}(x, ...)
}
\arguments{
\item{x}{A forecast_result object}
}
\description{
Plot method for forecast_result
}
