% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plot Forecast Data}
\usage{
plot_forecast(forecast_df)
}
\arguments{
\item{forecast_df}{Data frame from get_forecast}
}
\value{
A ggplot object
}
\description{
Plot Forecast Data
}
